FINAL_Travel_Packer/app/backend/requirements.txt
fastapi==0.108.0
uvicorn==0.29.0
transformers==4.46.3
torch==2.5.1
pillow==11.0.0
httpx==0.27.2
python-dotenv==1.0.1
scipy